FileCVS_app/requirements.txt
anyio==3.6.2
certifi==2022.12.7
chardet==5.1.0
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
orjson==3.8.9
Pillow==9.5.0
pydantic==1.10.7
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
rfc3986==1.5.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
Unidecode==1.3.6
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
